+++ frontend/src/App.js
@@ -1,13 +1,18 @@
 import { Messages } from './components/Messages'
+import { post } from './api/requests'
 
 
 // This line is not needed yet, but it will be needed later
 const backend = process.env.REACT_APP_BACKEND
+const endpoint = `${backend}/chat`
 
 
 const App = () => {
   return (
-    <Messages />
+    <Messages
+      endpoint={endpoint}
+      post={post}
+    />
   );
 }